Some Changes!

+++ main/Compiler/SymbolTable.h
@@ -1,33 +1,29 @@
 #pragma once
+#include <cstdint>
 #include <string>
-#include <vector>
 #include <unordered_map>
+#include <vector>
 
 class SymbolTable {
 private:
     std::unordered_map<std::string, size_t> nameToIndex;
-    std::vector<double> values;
+    std::vector<int32_t> values;
+
 public:
-    SymbolTable();                                   // defined in .cpp
+    SymbolTable();
     ~SymbolTable() = default;
 
-    // Returns index, creates if not exists
     size_t getIndex(const std::string& name);
 
-    // Check existence
     bool hasSymbol(const std::string& name) const;
     void addSymbol(const std::string& name);
 
-    // Get value by name (returns true if found)
-    bool getValue(const std::string& name, double& out) const;
-    // Set value by name
-    bool setValue(const std::string& name, double val);
+    bool getValue(const std::string& name, int32_t& out) const;
+    bool setValue(const std::string& name, int32_t val);
 
-    // For VM access
-    std::vector<double>& getValuesVector();
-    const std::vector<double>& getValuesVector() const;
+    std::vector<int32_t>& getValuesVector();
+    const std::vector<int32_t>& getValuesVector() const;
 
-    // Index-based access
-    double getValue(size_t idx) const;
-    void setValue(size_t idx, double val);
+    int32_t getValue(size_t idx) const;
+    void setValue(size_t idx, int32_t val);
 };
